update customer history page
